Fix bounds checks, integer overflows, and resource leaks across codebase
x224: validate LI byte against packet length before slicing

sec/cert: use uint64 arithmetic to prevent uint32 overflow in length check

client: guard empty drdynvc PDU before accessing data[0] in log statement

client: add || c.reconnecting to Close() to prevent double-close of done channel

client: track global VC reassembly memory with 256MB cap across all channels

websocket: cap WebSocket payload to 64MB before uint64-to-int cast

dispatcher/web: overflow-safe wsBuf allocation, dimension and data length

validation in OnStridedBitmap and bitmap send loops

broker: dimension validation and len(u.Data) >= needed check before bitmap copy

pixconv: validate len(src) and len(dst) at top of ConvertToRGBA

scard_pcsclite: bounds check at top of parseReaderState

dispatcher/web: 5s write deadline on disconnect notification to prevent

goroutine blocking on stalled connections

framebuffer: validate len(src) and srcStride in WriteRectStridedTopDown

rdpsnd/adpcm: cap IMA stereo interleave count to remaining samples, add

bounds checks before initial sample writes in both IMA and MS-ADPCM decoders

+++ protocol/sec/cert.go
@@ -52,10 +52,10 @@ func DecodeServerSecurityBlob(data []byte) (*ServerSecurityBlob, error) {
 		return nil, fmt.Errorf("server security blob too short: %d bytes", len(data))
 	}
 
-	serverRandomLen := binary.LittleEndian.Uint32(data[0:4])
-	serverCertLen := binary.LittleEndian.Uint32(data[4:8])
+	serverRandomLen := uint64(binary.LittleEndian.Uint32(data[0:4]))
+	serverCertLen := uint64(binary.LittleEndian.Uint32(data[4:8]))
 
-	if uint32(len(data)) < 8+serverRandomLen+serverCertLen {
+	if uint64(len(data)) < 8+serverRandomLen+serverCertLen {
 		return nil, fmt.Errorf("server security blob truncated: need %d, have %d",
 			8+serverRandomLen+serverCertLen, len(data))
 	}
